Moved Commands
Moved the `poll`, `raffle`, and `avatar` command to the `FUN` category.
