+++ src/main/java/com/beanbeanjuice/command/fun/AddPollCommand.java
@@ -1,4 +1,4 @@
-package com.beanbeanjuice.command.interaction;
+package com.beanbeanjuice.command.fun;
 
 import com.beanbeanjuice.main.CafeBot;
 import com.beanbeanjuice.utility.command.CommandContext;
@@ -176,6 +176,6 @@ public class AddPollCommand implements ICommand {
 
     @Override
     public CategoryType getCategoryType() {
-        return CategoryType.INTERACTION;
+        return CategoryType.FUN;
     }
 }
