+++ src/main/java/com/beanbeanjuice/command/fun/AvatarCommand.java
@@ -1,4 +1,4 @@
-package com.beanbeanjuice.command.interaction;
+package com.beanbeanjuice.command.fun;
 
 import com.beanbeanjuice.main.CafeBot;
 import com.beanbeanjuice.utility.command.CommandContext;
@@ -72,6 +72,6 @@ public class AvatarCommand implements ICommand {
 
     @Override
     public CategoryType getCategoryType() {
-        return CategoryType.INTERACTION;
+        return CategoryType.FUN;
     }
 }
