+++ src/main/java/com/beanbeanjuice/main/CafeBot.java
@@ -6,9 +6,9 @@ import com.beanbeanjuice.command.cafe.OrderCommand;
 import com.beanbeanjuice.command.cafe.ServeCommand;
 import com.beanbeanjuice.command.fun.MemeCommand;
 import com.beanbeanjuice.command.fun.JokeCommand;
-import com.beanbeanjuice.command.interaction.AddPollCommand;
-import com.beanbeanjuice.command.interaction.AddRaffleCommand;
-import com.beanbeanjuice.command.interaction.AvatarCommand;
+import com.beanbeanjuice.command.fun.AddPollCommand;
+import com.beanbeanjuice.command.fun.AddRaffleCommand;
+import com.beanbeanjuice.command.fun.AvatarCommand;
 import com.beanbeanjuice.command.moderation.SetCountingChannelCommand;
 import com.beanbeanjuice.command.generic.BugReportCommand;
 import com.beanbeanjuice.command.generic.FeatureRequestCommand;
